+++ src/component/GameMatch/GameMatch.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { ServiceTitle } from '../Main/ServiceTitle'
+
+export const GameMatch = () => {
+	return (
+		<div className='flex'>
+		<ServiceTitle title="Game" />
+	</div>
+	)
+}
