--- src/component/User/UserList.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const UserList = () => {
-	return (
-		<div>UserList</div>
-	)
-}
