+++ src/component/UserList/UserList.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { ServiceTitle } from '../Main/ServiceTitle'
+
+export const UserList = () => {
+	return (
+		<div className='flex'>
+			<ServiceTitle title="Friends" />
+		</div>
+	)
+}
